requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
black==23.10.1
boto3==1.33.11
botocore==1.33.11
Brotli==1.1.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.7
defusedxml==0.7.1
dj-database-url==2.1.0
Django==4.2.6
django-allauth==0.58.2
django-bootstrap-v5==1.0.11
django-storages==1.14.2
djangorestframework==3.14.0
Faker==19.12.0
gunicorn==21.2.0
idna==3.4
jmespath==1.0.1
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.2
pathspec==0.11.2
Pillow==10.1.0
platformdirs==3.11.0
psycopg==3.1.12
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
s3transfer==0.8.2
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.8.0
urllib3==2.0.7
whitenoise==6.6.0
